refactor(install cmd): extract logic into their own functions
